asdfghjkl.js: Remove `let` block
`let` blocks of the form

    let (var1 [= value1] [, var2 [= value2]] [, ..., varN [= valueN]]) block;

are non-standard, as specified in the following sections of the Mozilla
Developer Network page on `let`:

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Statements/let#Non-standard_let_extensions
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Statements/let#let_blocks

Removing the `let` block here for forward compatibility.

This change also gets rid of the `self` variable created inside the
`let` block, using an arrow function instead to obviate the need to save
a `this` reference.